requirements.txt
arxiv
openai
supabase
langchain-community
pypdf
python-dotenv
tiktoken
fastapi
sqlalchemy
uvicorn
asyncpg
